pyproject.toml
[tool.poetry]
name = "git-ai-commit"
version = "0.1.0"
description = "AI-powered Git commit message generator"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
openai = "^1.12.0"
pyyaml = "^6.0.1"
click = "^8.1.7"
gitpython = "^3.1.42"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gcommit = "src.cli:main"